Safari-611.1.10/LayoutTests/fast/css/display-compact-ignored.html

<script src="../../resources/js-test-pre.js"></script>
<span id=test style="display:compact"></span>
<script>
shouldBeEqualToString("getComputedStyle(test).display", "inline");
</script>
<script src="../../resources/js-test-post.js"></script>
