Safari-611.1.10/LayoutTests/fast/css/display-compact-ignored-expected.txt

PASS getComputedStyle(test).display is "inline"
PASS successfullyParsed is true

TEST COMPLETE
